Render initials color via inline CSS var so it actually shows
The per-user color was applied as a Tailwind `bg-(--x-9)` class built in
@posthog/core, but the app's Tailwind content globs only scan packages/ui
and apps/*, not packages/core — so those utilities were never generated and
every bubble fell back to the neutral gray default.

Return a CSS variable (var(--x-9)) instead and apply it via inline style on
the AvatarFallback, which doesn't depend on Tailwind class generation. Revert
the reviewer-avatar helper to its original standalone form (it has the same
latent scanning limitation but is out of scope here).

Generated-By: PostHog Code
Task-Id: 03a1a20f-7597-4267-bf73-0ef6224e71e7
