+++ packages/core/src/auth/avatarColor.ts
@@ -1,20 +1,18 @@
-// Deterministic per-user avatar color. A stable seed (user uuid, email, or name)
-// always maps to the same palette entry, so a given person keeps one color across
-// every view and session. All entries use a Radix "9" step whose accessible pairing
-// is white text (the light-9 scales — amber/yellow/lime/mint/sky — are excluded).
-const AVATAR_PALETTE = [
-  "bg-(--orange-9) text-white",
-  "bg-(--blue-9) text-white",
-  "bg-(--purple-9) text-white",
-  "bg-(--green-9) text-white",
-  "bg-(--pink-9) text-white",
-  "bg-(--teal-9) text-white",
-  "bg-(--red-9) text-white",
-  "bg-(--indigo-9) text-white",
-  "bg-(--cyan-9) text-white",
-  "bg-(--violet-9) text-white",
-  "bg-(--jade-9) text-white",
-  "bg-(--crimson-9) text-white",
+// Radix color scales whose "9" step pairs with white text (the light-9 scales —
+// amber/yellow/lime/mint/sky — are excluded so initials stay legible).
+const AVATAR_COLORS = [
+  "orange",
+  "blue",
+  "purple",
+  "green",
+  "pink",
+  "teal",
+  "red",
+  "indigo",
+  "cyan",
+  "violet",
+  "jade",
+  "crimson",
 ] as const;
 
 export function avatarColorSeedHash(seed: string): number {
@@ -25,6 +23,12 @@ export function avatarColorSeedHash(seed: string): number {
   return hash;
 }
 
-export function avatarColorClass(seed: string): string {
-  return AVATAR_PALETTE[avatarColorSeedHash(seed) % AVATAR_PALETTE.length];
+// Returns a CSS custom-property reference (e.g. "var(--orange-9)") for a stable
+// seed, applied via inline style. We return a CSS var rather than a Tailwind
+// `bg-*` class because this file lives in @posthog/core, which the app's Tailwind
+// content globs don't scan — a dynamically-chosen utility class here would never
+// be generated, so the bubble would fall back to the neutral default.
+export function avatarColorVar(seed: string): string {
+  const color = AVATAR_COLORS[avatarColorSeedHash(seed) % AVATAR_COLORS.length];
+  return `var(--${color}-9)`;
 }
